Make phrasematchsubq objects wrap inner cache objects instead of js cache objects to tear out the rest of the JS dependency in coalesce, and then move all the JS and uv/thread stuff out of coalesce.* and into binding.*
